Refactor(rendering): Split the template formatters
formatters.py held 512 lines, past the 400-line reviewability limit
the aislop code-quality gate applies.

Split it into a package at the same dotted path, grouped by what each
helper formats: numeric (counts, LOC, percentages, bytes), dates (ages
and timestamps), text (slugs, truncation, lists, feature names) and
filters (get_template_filters, the one place that needs all three).
The package __init__ re-exports the previous names, so the Jinja
filter registration and every import site are unchanged.
